# src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: JAWS API Document
  description: JAWS authentication and test-session API.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local backend, Vite proxy target
  - url: http://localhost:8081
    description: Local backend, frontend direct test/SSE target
tags:
  - name: Auth
    description: Authentication API
  - name: Test
    description: Test session API
paths:
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/test/start:
    post:
      tags: [Test]
      summary: Start a test session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestStartRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestStartResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/test/{sessionId}/stream:
    get:
      tags: [Test]
      summary: Subscribe to live session events with SSE
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Server-sent event stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/StreamEvent'
        '404':
          description: Session not found
  /api/test/{sessionId}/progress:
    get:
      tags: [Test]
      summary: Get session progress
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestProgressResponse'
  /api/test/{sessionId}/logs:
    get:
      tags: [Test]
      summary: Get action logs
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestLogsResponse'
  /api/test/{sessionId}/issues:
    get:
      tags: [Test]
      summary: Get detected issues
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestIssuesResponse'
  /api/test/{sessionId}/report:
    post:
      tags: [Test]
      summary: Create or get a report record
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestReportResponse'
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      example: 550e8400-e29b-41d4-a716-446655440000
  schemas:
    SignupRequest:
      type: object
      required: [userName, email, password]
      properties:
        userName:
          type: string
          example: tester
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: Password123!
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: Password123!
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        userId:
          type: integer
          format: int64
        userName:
          type: string
        role:
          type: string
          example: USER
    TestStartRequest:
      type: object
      required: [targetUrl]
      properties:
        targetUrl:
          type: string
          example: https://example.com
        agentConfig:
          type: object
          additionalProperties: true
          example: {}
    TestStartResponse:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
          example: READY
    TestProgressResponse:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
          example: RUNNING
        progress:
          type: integer
          example: 60
    TestLogsResponse:
      type: object
      properties:
        sessionId:
          type: string
        logs:
          type: array
          items:
            type: string
    TestIssuesResponse:
      type: object
      properties:
        sessionId:
          type: string
        issues:
          type: array
          items:
            type: string
    TestReportResponse:
      type: object
      properties:
        sessionId:
          type: string
        reportUrl:
          type: string
          example: /reports/550e8400-e29b-41d4-a716-446655440000.pdf
    StreamEvent:
      type: object
      properties:
        type:
          type: string
          enum: [progress, log, issue, status, complete]
        label:
          type: string
          example: Network
        message:
          type: string
          example: Network response timing checked
        progress:
          type: integer
          example: 80
        status:
          type: string
          enum: [running, paused, completed, failed]
        issueType:
          type: string
          enum: [error, warning]
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: URL is required.